Move duplicate into ReadOnlyCache
