+++ cache/src/main/kotlin/world/gregs/voidps/cache/memory/cache/MemoryCache.kt
@@ -1,13 +1,10 @@
 package world.gregs.voidps.cache.memory.cache
 
-class MemoryCache(
-    val data: Array<Array<Array<ByteArray?>?>?>,
-    indices: IntArray,
-    archives: Array<IntArray?>,
-    fileCounts: Array<IntArray?>,
-    files: Array<Array<IntArray?>?>,
-    hashes: Map<Int, Int>?
-) : ReadOnlyCache(indices, archives, fileCounts, files, hashes) {
+
+class MemoryCache(indexCount: Int) : ReadOnlyCache(indexCount) {
+
+    lateinit var data: Array<Array<Array<ByteArray?>?>?>
+
 
     override fun data(index: Int, archive: Int, file: Int, xtea: IntArray?): ByteArray? {
         return data.getOrNull(index)?.getOrNull(archive)?.getOrNull(file)
